Move the helper functions to utils
Refs PV-491
