+++ src/utils.ts
@@ -1,5 +1,5 @@
-import { format } from 'date-fns';
-import { ParkingContractType, Permit } from './types';
+import { format, intervalToDuration } from 'date-fns';
+import { ParkingContractType, Permit, Product } from './types';
 
 export const validateTime = (time: string): boolean =>
   /^([0-1]?[0-9]|2[0-4]):([0-5][0-9])(:[0-5][0-9])?$/.test(time);
@@ -42,3 +42,29 @@ export const isOpenEndedPermitStarted = (
       p.contractType === ParkingContractType.OPEN_ENDED &&
       new Date(p.startTime as string).valueOf() < new Date().valueOf()
   );
+
+export const dateAsNumber = (date: Date | string): number =>
+  new Date(date).valueOf();
+
+export const getMonthCount = (
+  permitEndStartDate: Date,
+  permitStartTime: string,
+  product: Product
+): number => {
+  // It should consider the start time of the permit.
+  // Eg: If permit was bought and started just before the permitEndStartDate
+  // then it should be counted as a full month used and if the start date is in
+  // the future. It should refund the whole month.
+  if (
+    dateAsNumber(permitStartTime) > permitEndStartDate.valueOf() ||
+    dateAsNumber(product.startDate) > permitEndStartDate.valueOf()
+  ) {
+    return product.quantity;
+  }
+
+  const intervalDuration = intervalToDuration({
+    start: new Date(),
+    end: new Date(product.endDate),
+  });
+  return intervalDuration.months || 0;
+};
